add:增加useInitEditorHooks，初始化编辑器
