+++ app/renderer/src/main/src/components/MilkdownEditor/MilkdownEditor.tsx
@@ -1,44 +1,14 @@
-import {defaultValueCtx, Editor, editorViewOptionsCtx, rootCtx} from "@milkdown/kit/core"
+import {defaultValueCtx} from "@milkdown/kit/core"
 import React, {useEffect, useRef, useState} from "react"
 
-import {Milkdown, MilkdownProvider, useEditor} from "@milkdown/react"
-import {
-    blockquoteSchema,
-    codeBlockSchema,
-    commonmark,
-    hrSchema,
-    listItemSchema,
-    syncHeadingIdPlugin
-} from "@milkdown/kit/preset/commonmark"
-import {gfm} from "@milkdown/kit/preset/gfm"
-import {history} from "@milkdown/kit/plugin/history"
-import {clipboard} from "@milkdown/kit/plugin/clipboard"
+import {Milkdown, MilkdownProvider} from "@milkdown/react"
 
-import {ProsemirrorAdapterProvider, useNodeViewFactory, usePluginViewFactory} from "@prosemirror-adapter/react"
-import {tooltip, TooltipView} from "./Tooltip/Tooltip"
-import {BlockView} from "./Block/Block"
-import {block, blockConfig} from "@milkdown/plugin-block" // 引入block插件
-import {cursor} from "@milkdown/kit/plugin/cursor"
-import {imageBlockComponent, imageBlockConfig} from "@milkdown/kit/component/image-block"
-
-import {linkTooltipPlugin, linkTooltipConfig} from "@milkdown/kit/component/link-tooltip"
-import {listener, listenerCtx} from "@milkdown/kit/plugin/listener"
+import {ProsemirrorAdapterProvider} from "@prosemirror-adapter/react"
 
 import "./css/index.scss"
-import {yakitInfo, yakitNotify} from "@/utils/notification"
-import {placeholderConfig, placeholderPlugin} from "./Placeholder"
-import {$view, getMarkdown} from "@milkdown/kit/utils"
-import {CustomCodeComponent} from "./CodeBlock/CodeBlock"
-import {Blockquote} from "./Blockquote"
-import {
-    CollabStatus,
-    CustomMilkdownProps,
-    DeleteOSSFileItem,
-    MilkdownCollabProps,
-    MilkdownEditorProps
-} from "./MilkdownEditorType"
-import {alterCustomPlugin} from "./utils/alertPlugin"
-import {commentCustomPlugin} from "./utils/commentPlugin"
+import {yakitNotify} from "@/utils/notification"
+import {getMarkdown} from "@milkdown/kit/utils"
+import {CollabStatus, CustomMilkdownProps, DeleteOSSFileItem, MilkdownEditorProps} from "./MilkdownEditorType"
 
 import {diffLines} from "diff"
 import {
@@ -50,35 +20,15 @@ import {
     useThrottleFn,
     useUpdateEffect
 } from "ahooks"
-import {underlineCustomPlugin} from "./utils/underline"
-import {ListItem} from "./ListItem/ListItem"
-import {Ctx, MilkdownPlugin} from "@milkdown/kit/ctx"
-import {trailing} from "@milkdown/kit/plugin/trailing"
 
-import {upload, uploadConfig} from "@milkdown/kit/plugin/upload"
-import type {Node} from "@milkdown/kit/prose/model"
-import {imageInlineComponent, inlineImageConfig} from "@milkdown/kit/component/image-inline"
-
-import {html} from "atomico"
-import {fileCustomSchema, uploadCustomPlugin} from "./utils/uploadPlugin"
-import {CustomFile} from "./CustomFile/CustomFile"
-import {insertImageBlockCommand} from "./utils/imageBlock"
-
-import {listCustomPlugin} from "./utils/listPlugin"
-import {headingCustomPlugin} from "./utils/headingPlugin"
-import {codeCustomPlugin} from "./utils/codePlugin"
-import {MilkdownHr} from "./MilkdownHr/MilkdownHr"
-
-import {tableBlock} from "@milkdown/kit/component/table-block"
-import {ImgMaxSize} from "@/pages/pluginEditor/pluginImageTextarea/PluginImageTextarea"
-import {httpDeleteOSSResource, httpUploadImgBase64} from "@/apiUtils/http"
-import {deletedFileUrlsCtx, trackDeletePlugin} from "./utils/trackDeletePlugin"
+import {httpDeleteOSSResource} from "@/apiUtils/http"
+import {deletedFileUrlsCtx} from "./utils/trackDeletePlugin"
 import moment from "moment"
 import {useStore} from "@/store"
 import {CollabManager, CollabUserInfo} from "./CollabManager"
 import emiter from "@/utils/eventBus/eventBus"
 
-import {collab, collabServiceCtx} from "@milkdown/plugin-collab"
+import {collabServiceCtx} from "@milkdown/plugin-collab"
 import {showYakitModal} from "../yakitUI/YakitModal/YakitModalConfirm"
 import {tokenOverdue} from "@/services/fetch"
 import {isBoolean} from "lodash"
@@ -86,7 +36,7 @@ import {notepadSaveStatus} from "./WebsocketProvider/constants"
 import {toAddNotepad, toEditNotepad} from "@/pages/notepadManage/notepadManage/NotepadManage"
 import {API} from "@/services/swagger/resposeType"
 import {apiSaveNotepad} from "@/pages/notepadManage/notepadManage/utils"
-
+import useInitEditorHooks, {InitEditorHooksCollabProps} from "./utils/initEditor"
 const markdown1 = `
 
 1. 5565
@@ -104,7 +54,7 @@ Maybe more? ![]()
 const markdown2 = `#ggg
 fsdfsdf
 `
-const randomUserColor = () => `#${Math.floor(Math.random() * 16777215).toString(16)}`
+export const randomMilkdownUserColor = () => `#${Math.floor(Math.random() * 16777215).toString(16)}`
 
 const saveHistoryIntervalTime = 1000 * 60
 
@@ -124,66 +74,13 @@ export const isDifferenceGreaterThan30Seconds = (timestamp1, timestamp2) => {
     return difference > 1000 * 30
 }
 const CustomMilkdown: React.FC<CustomMilkdownProps> = React.memo((props) => {
-    const {
-        type,
-        readonly,
-        defaultValue,
-        collabProps,
-        setEditor,
-        customPlugin,
-        onMarkdownUpdated,
-        onSaveContentBeforeDestroy
-    } = props
+    const userInfo = useStore((s) => s.userInfo)
+
+    const {defaultValue, collabProps, setEditor, onSaveContentBeforeDestroy} = props
 
     const milkdownRef = useRef<HTMLDivElement>(null)
     const [inViewport = true] = useInViewport(milkdownRef)
-
-    const nodeViewFactory = useNodeViewFactory()
-    const pluginViewFactory = usePluginViewFactory()
-
-    const userInfo = useStore((s) => s.userInfo)
-
-    const collabParams: MilkdownCollabProps = useCreation(() => {
-        if (!collabProps) {
-            const def: MilkdownCollabProps = {
-                title: "",
-                enableCollab: false,
-                milkdownHash: "",
-                routeInfo: {
-                    pageId: "",
-                    route: null
-                },
-                onChangeWSLinkStatus: () => {},
-                onChangeOnlineUser: () => {},
-                onSetTitle: () => {}
-            }
-            return def
-        }
-        return collabProps
-    }, [collabProps])
-
-    const uploadImg = useMemoizedFn(async (image) => {
-        if (image.size > ImgMaxSize) {
-            yakitNotify("error", "图片大小不能超过1M")
-            return ""
-        }
-        try {
-            const base64 = await getBase64(image)
-            const src = await httpUploadImgBase64({
-                base64,
-                imgInfo: {
-                    filename: image.name || "image.png",
-                    contentType: image.type || "image/png"
-                },
-                type: collabParams.enableCollab ? "notepad" : "img",
-                filedHash: collabParams.milkdownHash
-            })
-            return src
-        } catch (error) {
-            return ""
-        }
-    })
-
+    const collabManagerRef = useRef<CollabManager>() // 协作管理器
     useEffect(() => {
         return () => {
             // 统一
@@ -195,266 +92,34 @@ const CustomMilkdown: React.FC<CustomMilkdownProps> = React.memo((props) => {
     }, [])
 
     //#region 编辑器初始
-    const {get, loading} = useEditor(
-        (root) => {
-            //#region 编辑器引用的相关插件 start
-            const blockPlugins = [
-                block,
-                (ctx: Ctx) => () => {
-                    ctx.set(block.key, {
-                        view: pluginViewFactory({
-                            component: () => <BlockView type={type} notepadHash={collabParams?.milkdownHash} />
-                        })
-                    })
+    const collabParams: InitEditorHooksCollabProps = useCreation(() => {
+        if (!collabProps) {
+            const def: InitEditorHooksCollabProps = {
+                title: "",
+                enableCollab: false,
+                milkdownHash: "",
+                routeInfo: {
+                    pageId: "",
+                    route: null
                 },
-                (ctx: Ctx) => () => {
-                    ctx.update(blockConfig.key, () => ({
-                        filterNodes: (pos, node) => {
-                            if (node.type.name === "paragraph" && !node.content.size) {
-                                return true
-                            }
-                            return false
-                        }
-                    }))
-                }
-            ].flat()
-            const placeholder = [placeholderConfig, placeholderPlugin]
-            const uploadPlugins = [
-                ...uploadCustomPlugin(),
-                upload,
-                $view(fileCustomSchema.node, () =>
-                    nodeViewFactory({
-                        component: () => <CustomFile type={type} />
-                    })
-                ),
-                (ctx: Ctx) => () => {
-                    ctx.update(uploadConfig.key, (prev) => ({
-                        ...prev,
-                        uploader: async (files, schema) => {
-                            const images: File[] = []
-                            for (let i = 0; i < files.length; i++) {
-                                const file = files.item(i)
-                                if (!file) {
-                                    continue
-                                }
-
-                                // You can handle whatever the file type you want, we handle image here.
-                                if (!file.type.includes("image")) {
-                                    continue
-                                }
-                                if (file.size > ImgMaxSize) {
-                                    yakitNotify("error", "图片大小不能超过1M")
-                                    continue
-                                }
-                                images.push(file)
-                            }
-                            const nodes: Node[] = await Promise.all(
-                                images.map(async (image) => {
-                                    const alt = image.name
-                                    try {
-                                        const src = await uploadImg(image)
-                                        return schema.nodes["image-block"].createAndFill({
-                                            src,
-                                            alt
-                                        }) as Node
-                                    } catch (error) {
-                                        return schema.nodes.image.createAndFill({
-                                            src: "",
-                                            alt
-                                        }) as Node
-                                    }
-                                })
-                            )
-                            return nodes
-                        }
-                    }))
-                }
-            ].flat()
-
-            const imagePlugin = [
-                imageBlockComponent,
-                imageInlineComponent,
-                insertImageBlockCommand,
-                (ctx: Ctx) => () => {
-                    ctx.update(imageBlockConfig.key, (value) => ({
-                        ...value,
-                        captionIcon: () => html`
-                            <svg
-                                xmlns="http://www.w3.org/2000/svg"
-                                width="24"
-                                height="24"
-                                viewBox="0 0 24 24"
-                                fill="none"
-                            >
-                                <path
-                                    d="M7 8H17M7 12H11M12 20L8 16H5C3.89543 16 3 15.1046 3 14V6C3 4.89543 3.89543 4 5 4H19C20.1046 4 21 4.89543 21 6V14C21 15.1046 20.1046 16 19 16H16L12 20Z"
-                                    stroke="currentColor"
-                                    stroke-width="1.5"
-                                    stroke-linecap="round"
-                                    stroke-linejoin="round"
-                                />
-                            </svg>
-                        `,
-                        onUpload: async (image: File) => {
-                            const url = uploadImg(image)
-                            return url
-                        }
-                    }))
-                },
-                (ctx: Ctx) => () => {
-                    ctx.update(inlineImageConfig.key, (value) => ({
-                        ...value,
-                        imageIcon: () =>
-                            html`<svg
-                                width="24"
-                                height="24"
-                                viewBox="0 0 24 24"
-                                fill="none"
-                                xmlns="http://www.w3.org/2000/svg"
-                            >
-                                <path
-                                    d="M4 16L8.58579 11.4142C9.36683 10.6332 10.6332 10.6332 11.4142 11.4142L16 16M14 14L15.5858 12.4142C16.3668 11.6332 17.6332 11.6332 18.4142 12.4142L20 14M14 8H14.01M6 20H18C19.1046 20 20 19.1046 20 18V6C20 4.89543 19.1046 4 18 4H6C4.89543 4 4 4.89543 4 6V18C4 19.1046 4.89543 20 6 20Z"
-                                    stroke="currentColor"
-                                    stroke-width="1.5"
-                                    stroke-linecap="round"
-                                    stroke-linejoin="round"
-                                />
-                                <circle cx="14" cy="8" r="1" fill="currentColor" />
-                            </svg> `,
-                        onUpload: async (image: File) => {
-                            const url = uploadImg(image)
-                            return url
-                        }
-                    }))
-                }
-            ].flat()
-
-            const linkTooltip = [
-                linkTooltipPlugin,
-                (ctx: Ctx) => () => {
-                    ctx.update(linkTooltipConfig.key, (defaultConfig) => ({
-                        ...defaultConfig,
-                        linkIcon: () => "🔗",
-                        editButton: () => "✎",
-                        removeButton: () => "❌",
-                        confirmButton: () => "✔️",
-                        onCopyLink: (link: string) => {
-                            yakitInfo("Link copied")
-                        }
-                    }))
-                }
-            ].flat()
-
-            const listPlugin = [
-                ...listCustomPlugin(),
-                $view(listItemSchema.node, () =>
-                    nodeViewFactory({
-                        component: ListItem
-                    })
-                )
-            ].flat()
-            const headingPlugin = [...headingCustomPlugin()].flat()
-            const codePlugin = [
-                ...codeCustomPlugin(),
-                $view(codeBlockSchema.node, () => {
-                    return nodeViewFactory({
-                        component: CustomCodeComponent,
-                        stopEvent: (e) => true
-                    })
-                })
-            ].flat()
-            const blockquotePlugin = [
-                $view(blockquoteSchema.node, () =>
-                    nodeViewFactory({
-                        component: Blockquote
-                    })
-                )
-            ].flat()
-
-            const alterPlugin = [...alterCustomPlugin()].flat()
-
-            const underlinePlugin = [...underlineCustomPlugin()].flat()
-
-            const commentPlugin = [...commentCustomPlugin()].flat()
-
-            const hrPlugin = [
-                $view(hrSchema.node, () =>
-                    nodeViewFactory({
-                        component: MilkdownHr
-                    })
-                )
-            ].flat()
-            //#endregion
-            return (
-                Editor.make()
-                    .config((ctx) => {
-                        ctx.set(rootCtx, root)
-                        ctx.set(tooltip.key, {
-                            view: pluginViewFactory({
-                                component: TooltipView
-                            })
-                        })
-                        // 配置为只读
-                        ctx.set(editorViewOptionsCtx, {
-                            editable: () => !readonly
-                        })
-                        ctx.set(defaultValueCtx, defaultValue || "")
-                        onCollab(ctx)
-                        const listener = ctx.get(listenerCtx)
-                        listener.markdownUpdated((ctx, nextMarkdown, prevMarkdown) => {
-                            const isSave = nextMarkdown !== prevMarkdown
-                            if (collabParams.enableCollab && isSave) {
-                                onSaveHistory(nextMarkdown)
-                            }
-                            onMarkdownUpdated && onMarkdownUpdated(nextMarkdown, prevMarkdown)
-                        })
-                    })
-                    .use(commonmark.filter((x) => x !== syncHeadingIdPlugin))
-                    .use(gfm)
-                    .use(cursor)
-                    .use(tooltip)
-                    .use(history)
-                    .use(clipboard)
-                    // trailing
-                    .use(trailing)
-                    // collab
-                    .use(collab)
-                    // listener
-                    .use(listener)
-                    // blockquote
-                    .use(blockquotePlugin)
-                    // block
-                    .use(blockPlugins)
-                    // upload
-                    .use(uploadPlugins)
-                    // image
-                    .use(imagePlugin)
-                    // listItem
-                    .use(listPlugin)
-                    .use(headingPlugin)
-                    // code
-                    .use(codePlugin)
-                    // linkTooltip
-                    .use(linkTooltip)
-                    // placeholder
-                    .use(placeholder)
-                    // table
-                    .use(tableBlock)
-                    // alterCustomPlugin
-                    .use(alterPlugin)
-                    // underlinePlugin
-                    .use(underlinePlugin)
-                    // commentPlugin
-                    .use(commentPlugin)
-                    // hrPlugin
-                    .use(hrPlugin)
-                    // trackDeletePlugin
-                    .use(trackDeletePlugin())
-                    .use(customPlugin || [])
-            )
-        },
-        [readonly, defaultValue, type, collabParams.enableCollab, collabParams.milkdownHash]
-    )
+                onChangeWSLinkStatus: () => {},
+                onChangeOnlineUser: () => {},
+                onSetTitle: () => {},
+                onCollab: () => {},
+                onSaveHistory: () => {}
+            }
+            return def
+        }
+        return {
+            ...collabProps,
+            onCollab: (ctx) => onCollab(ctx),
+            onSaveHistory: (newValue) => onSaveHistory(newValue)
+        }
+    }, [collabProps])
+    const {get, loading} = useInitEditorHooks({
+        ...props,
+        collabProps: collabParams
+    })
     /**更新最新的editor */
     useEffect(() => {
         if (loading) return
@@ -524,7 +189,6 @@ const CustomMilkdown: React.FC<CustomMilkdownProps> = React.memo((props) => {
     //#endregion 删除资源 end
 
     //#region 协作文档
-    const collabManagerRef = useRef<CollabManager>()
     const wsStatusRef = useRef<CollabStatus>({
         status: "disconnected",
         isSynced: false,
@@ -561,7 +225,7 @@ const CustomMilkdown: React.FC<CustomMilkdownProps> = React.memo((props) => {
         const user = {
             userId: userInfo.user_id || 0,
             name: userInfo.companyName || "",
-            color: `${randomUserColor()}`,
+            color: `${randomMilkdownUserColor()}`,
             heardImg: userInfo.companyHeadImg || ""
         }
         const collabService = ctx.get(collabServiceCtx)
@@ -575,6 +239,7 @@ const CustomMilkdown: React.FC<CustomMilkdownProps> = React.memo((props) => {
         collabManagerRef.current.on("online-users", onSetOnlineUsers)
         collabManagerRef.current.on("link-status-onchange", onLineStatus)
         collabManagerRef.current.on("sync-title", onSetTitle)
+        collabManagerRef.current.on("update-diff", onUpdateDiff)
     })
 
     /**同步标题 */
@@ -583,6 +248,29 @@ const CustomMilkdown: React.FC<CustomMilkdownProps> = React.memo((props) => {
             collabParams?.onSetTitle(value)
         }),
         {wait: 200, leading: true}
+    ).run
+
+    /*TODO - 保存历史暂时在CollabManager内部，后续将content也要保存在历史中，确保容错*/
+    const onUpdateDiff = useDebounceFn(
+        useMemoizedFn((val) => {
+            // const content = get()?.action(getMarkdown())
+            // const item = {
+            //     ...val,
+            //     content
+            // }
+            // const value = sessionStorage.getItem("versions")
+            // try {
+            //     if (value) {
+            //         const versionList = value ? JSON.parse(value) : []
+            //         versionList.push(item)
+            //         sessionStorage.setItem("versions", JSON.stringify(versionList))
+            //     } else {
+            //         sessionStorage.setItem("versions", JSON.stringify([item]))
+            //     }
+            // } catch (error) {
+            // }
+        }),
+        {wait: 1000}
     ).run
 
     /**在线链接状态 */
@@ -656,6 +344,7 @@ const CustomMilkdown: React.FC<CustomMilkdownProps> = React.memo((props) => {
         const s = showYakitModal({
             title: "文档不存在/已经被删除",
             content: <span>错误原因:{event.reason}</span>,
+            maskClosable: false,
             closable: false,
             showConfirmLoading: true,
             onOkText: "保存当前文档",
@@ -695,6 +384,7 @@ const CustomMilkdown: React.FC<CustomMilkdownProps> = React.memo((props) => {
                     错误原因:{event.code}:{event.reason}
                 </span>
             ),
+            maskClosable: false,
             closable: false,
             onOkText: "关闭页面",
             cancelButtonProps: {style: {display: "none"}},
